Refactor BarNavigation and SectionHeader components
