+++ src/components/BarNavigation.tsx
@@ -12,26 +12,20 @@ import Menu from "@mui/material/Menu";
 import MenuItem from "@mui/material/MenuItem";
 import ChevronLeftIcon from "@mui/icons-material/ChevronLeft";
 import ChevronRightIcon from "@mui/icons-material/ChevronRight";
-import DashboardIcon from "@mui/icons-material/Dashboard";
-import PersonIcon from "@mui/icons-material/Person";
-import PeopleIcon from "@mui/icons-material/People";
-import DriveEtaIcon from "@mui/icons-material/DriveEta";
-import TimeToLeaveIcon from "@mui/icons-material/TimeToLeave";
-import ArticleIcon from "@mui/icons-material/Article";
-import SettingsIcon from "@mui/icons-material/Settings";
+
 import AccountCircleIcon from "@mui/icons-material/AccountCircle";
 import Badge from "@mui/material/Badge";
 import NotificationsActiveIcon from "@mui/icons-material/NotificationsActive";
 import Popover from "@mui/material/Popover";
 import List from "@mui/material/List";
 import { ListItemButton, ListItemText, Stack } from "@mui/material";
-import type { SectionHeaderProps } from "./SectionHeader";
+import { sectionHeaders, handleSectionHeaderClick } from "./SectionHeader";
 import SectionHeader from "./SectionHeader";
 import { createBrowserHistory } from "history";
-// import { AppBarStyled } from "./AppBar";
 
 const drawerWidth = 240;
 const history = createBrowserHistory();
+
 const openedMixin = (theme: Theme): CSSObject => ({
 	width: drawerWidth,
 	transition: theme.transitions.create("width", {
@@ -112,83 +106,6 @@ export default function MiniDrawer() {
 		setOpen(false);
 		sectionHeaders[4].open = false;
 	};
-
-	const handleSectionHeaderClick = (sectionHeaderText: string) => {
-		history.push(`/${sectionHeaderText.toLowerCase()}`);
-		const sectionIndex = sectionHeaders.findIndex(
-			(header) => header.text === sectionHeaderText
-		);
-
-		if (sectionIndex !== -1) {
-			const updatedSectionHeaders = [...sectionHeaders];
-			updatedSectionHeaders[sectionIndex].open =
-				!updatedSectionHeaders[sectionIndex].open;
-			setSectionHeaders(updatedSectionHeaders);
-		}
-
-		if (sectionIndex === 4) {
-			setOpen(true);
-		} else setOpen(open);
-	};
-
-	const [sectionHeaders, setSectionHeaders] = React.useState<
-		SectionHeaderProps[]
-	>([
-		{
-			text: "Dashboard",
-			icon: <DashboardIcon />,
-			nestedItems: [],
-			open: false,
-			handleClick: () => {},
-		},
-		{
-			text: "Drivers",
-			icon: <PersonIcon />,
-			nestedItems: [],
-			open: false,
-			handleClick: () => {},
-		},
-		{
-			text: "Vehicles",
-			icon: <DriveEtaIcon />,
-			nestedItems: [],
-			open: false,
-			handleClick: () => {},
-		},
-		{
-			text: "Passengers",
-			icon: <PeopleIcon />,
-			nestedItems: [],
-			open: false,
-			handleClick: () => {},
-		},
-		{
-			text: "Reports",
-			icon: <ArticleIcon />,
-			nestedItems: [
-				"Accident/malfunction report",
-				"Driver details report",
-				"Passenger details report",
-				"Vehicle details report",
-				"Distance report",
-				"Filtered reports",
-			],
-			open: false, // Added this to match the expected type
-			handleClick: () => handleSectionHeaderClick("Reports"),
-		},
-		{
-			text: "Trips",
-			icon: <TimeToLeaveIcon />,
-			open: false,
-			handleClick: () => {},
-		},
-		{
-			text: "Settings",
-			icon: <SettingsIcon />,
-			open: false,
-			handleClick: () => {},
-		},
-	]);
 
 	const [anchorEl, setAnchorEl] = React.useState<HTMLElement | null>(null);
 
@@ -346,21 +263,6 @@ export default function MiniDrawer() {
 					))}
 				</List>
 			</Drawer>
-			{/* <Box
-				component="main"
-				sx={{ flexGrow: 1, p: 3 }}
-			>
-				{
-					<Router>
-						<Routes>
-							<Route
-								path="/dashboard"
-								element={<DashboardPage />}
-							/>
-						</Routes>
-					</Router>
-				}
-			</Box> */}
 		</Stack>
 	);
 }
